requirements.txt
requests
python-docx
openai
ffmpeg-python